+++ Collections/Buffer/BufferMgr.h
@@ -8,10 +8,7 @@ namespace Buffer {
     class BufferMgr: public Admin::InstanceMgr {
         public:
             template <typename T>
-            Buffer <T>* initBuffer (size_t instanceId, 
-                                    e_type type, 
-                                    size_t capacity) {
-
+            Buffer <T>* initBuffer (uint32_t instanceId, e_type type, size_t capacity) {
                 /* Create and add buffer object to pool
                 */
                 if (m_instancePool.find (instanceId) == m_instancePool.end()) {
