+++ Core/Cmds/VKCmdBuffer.h
@@ -10,7 +10,7 @@ namespace Renderer {
     class VKCmdBuffer: protected virtual VKDeviceMgr {
         private:
             static Log::Record* m_VKCmdBufferLog;
-            const size_t m_instanceId = g_collectionsId++;
+            const uint32_t m_instanceId = g_collectionsId++;
 
         public:
             VKCmdBuffer (void) {
