+++ Core/Pipeline/VKColorBlend.h
@@ -9,7 +9,7 @@ namespace Renderer {
     class VKColorBlend: protected virtual VKPipelineMgr {
         private:
             static Log::Record* m_VKColorBlendLog;
-            const size_t m_instanceId = g_collectionsId++;
+            const uint32_t m_instanceId = g_collectionsId++;
             
         public:
             VKColorBlend (void) {
